+++ src/main/java/com/kt/smartfarm/service/AuthCheckService.java
@@ -5,9 +5,9 @@ import com.kt.cmmn.util.AuthorityChecker;
 import java.util.List;
 
 public interface AuthCheckService {
-	Boolean authCheck(Integer gsmKey, Integer houseId);
+	Boolean authCheck(Long gsmKey, Long houseId);
 
-    Boolean authCheck(Integer gsmKey, Integer houseId, List<Integer> gsmKeyList, List<Integer> houseIdList);
+    Boolean authCheck(Long gsmKey, Long houseId, List<Long>gsmKeyList, List<Long>houseIdList);
 
     String getAuthUserIdx();
 	AuthorityChecker getAuthCheck();
