+++ src/main/java/com/kt/smartfarm/service/AutoSyncService.java
@@ -3,6 +3,5 @@ package com.kt.smartfarm.service;
 import javax.servlet.http.HttpServletRequest;
 
 public interface AutoSyncService {
-
-    Integer autoSync(Integer gmsKey, HttpServletRequest request);
+    Integer autoSync(Long gmsKey, HttpServletRequest request);
 }
