+++ src/main/java/com/kt/smartfarm/service/HouseEnvService.java
@@ -9,30 +9,30 @@ import org.springframework.stereotype.Component;
 @Component
 public interface HouseEnvService {	
 	HouseEnvVO insert( HouseEnvVO house);
-	Integer delete(Integer gsmKey, Integer greenHouseId);
+	Integer delete(Long gsmKey, Long greenHouseId);
 	HouseEnvVO update(HouseEnvVO house);
-	HashMap<String,Object>  get(Integer gsmKey, Integer greenHouseId, Boolean isSmartfarmSystem);
-	List<HashMap<String, Object>> list(Integer gsmKey, boolean all, boolean detail, Boolean isSmartfarmSystem);
-	List<HashMap<String, Object>> list(Integer gsmKey, boolean all, boolean detail, Boolean isSmartfarmSystem, Boolean isCCTVOnly);
+	HashMap<String,Object>  get(Long gsmKey, Long greenHouseId, Boolean isSmartfarmSystem);
+	List<HashMap<String, Object>> list(Long gsmKey, boolean all, boolean detail, Boolean isSmartfarmSystem);
+	List<HashMap<String, Object>> list(Long gsmKey, boolean all, boolean detail, Boolean isSmartfarmSystem, Boolean isCCTVOnly);
 	List<HashMap<String,Object>> selectHouseTypeList();
 	List<ProductVO> selectProductList();
 	HashMap<String,Object> insertHouseDeviceMap(HashMap<String, Object> map);
 	HashMap<String,Object> deleteHouseDeviceMap(HashMap<String, Object> map);
-	List<DeviceEnvVO> houseDeviceList(Integer houseId);
+	List<DeviceEnvVO> houseDeviceList(Long houseId);
 	List<HashMap<String, Object>> getAllList();
 	Integer insertForecastData(LinkedHashMap<String, Object> hm);
-	List<HashMap<String,Object>> getWeatherCast(Integer houseId, String fromDate, String toDate, Boolean isSmartfarmSystem);
+	List<HashMap<String,Object>> getWeatherCast(Long houseId, String fromDate, String toDate, Boolean isSmartfarmSystem);
 	List<HashMap<String,Object>> getWeatherCategory();
 	Integer insertSunriseData(HashMap<String, Object> hm);
 
-	List<HashMap<String,Object>> groundDeviceList(Integer houseId);
+	List<HashMap<String,Object>> groundDeviceList(Long houseId);
 
 	// CCTV
 	Integer insertCctv(CCTVSettingVO cctv);
 	Integer deleteCctv(Integer id);
 	Integer updateCctv(CCTVSettingVO cctv);
-	List<HashMap<String, Object>> getCctvsByHouseId(Integer houseId);
-	HashMap<String,Object> houseDeviceInfoList(Integer houseId);
+	List<HashMap<String, Object>> getCctvsByHouseId(Long houseId);
+	HashMap<String,Object> houseDeviceInfoList(Long houseId);
 
-	Integer copyToNewGSM(Integer fromGsmKey, Integer toGsmKey);
+	Integer copyToNewGSM(Long fromGsmKey, Long toGsmKey);
 }
